[Handshake] Fix library dependencies
