+++ include/circt/Conversion/HandshakeToHW.h
@@ -27,32 +27,8 @@ namespace circt {
 #define GEN_PASS_DECL_HANDSHAKETOHW
 #include "circt/Conversion/Passes.h.inc"
 
-namespace esi {
-class ChannelType;
-} // namespace esi
-
 std::unique_ptr<mlir::Pass> createHandshakeToHWPass();
 
-namespace handshake {
-
-// Attribute name for the name of a predeclaration of the to-be-lowered
-// hw.module from a handshake function.
-static constexpr const char *kPredeclarationAttr = "handshake.module_name";
-
-// Converts 't' into a valid HW type. This is strictly used for converting
-// 'index' types into a fixed-width type.
-Type toValidType(Type t);
-
-// Wraps a type into an ESI ChannelType type. The inner type is converted to
-// ensure comprehensability with the RTL dialects.
-esi::ChannelType esiWrapper(mlir::Type t);
-
-// Returns the hw::ModulePortInfo that corresponds to the given handshake
-// operation and its in- and output types.
-hw::ModulePortInfo getPortInfoForOpTypes(mlir::Operation *op, TypeRange inputs,
-                                         TypeRange outputs);
-
-} // namespace handshake
 } // namespace circt
 
 #endif // CIRCT_CONVERSION_HANDSHAKETOHW_H
