+++ include/circt/Dialect/Handshake/HandshakePasses.h
@@ -47,34 +47,6 @@ createHandshakeInsertBuffersPass(const std::string &strategy = "all",
 std::unique_ptr<mlir::Pass> createHandshakeLockFunctionsPass();
 std::unique_ptr<mlir::Pass> createHandshakeSplitMergesPass();
 
-/// Iterates over the handshake::FuncOp's in the program to build an instance
-/// graph. In doing so, we detect whether there are any cycles in this graph, as
-/// well as infer a top function for the design by performing a topological sort
-/// of the instance graph. The result of this sort is placed in sortedFuncs.
-using InstanceGraph = std::map<std::string, std::set<std::string>>;
-LogicalResult resolveInstanceGraph(ModuleOp moduleOp,
-                                   InstanceGraph &instanceGraph,
-                                   std::string &topLevel,
-                                   SmallVectorImpl<std::string> &sortedFuncs);
-
-// Checks all block arguments and values within op to ensure that all
-// values have exactly one use.
-LogicalResult verifyAllValuesHasOneUse(handshake::FuncOp op);
-
-// Adds sink operations to any unused value in r.
-LogicalResult addSinkOps(Region &r, OpBuilder &rewriter);
-
-// Adds fork operations to any value with multiple uses in r.
-LogicalResult addForkOps(Region &r, OpBuilder &rewriter);
-void insertFork(Value result, bool isLazy, OpBuilder &rewriter);
-
-// Adds a locking mechanism around the region.
-LogicalResult lockRegion(Region &r, OpBuilder &rewriter);
-
-// Applies the spcified buffering strategy on the region r.
-LogicalResult bufferRegion(Region &r, OpBuilder &rewriter, StringRef strategy,
-                           unsigned bufferSize);
-
 /// Generate the code for registering passes.
 #define GEN_PASS_REGISTRATION
 #include "circt/Dialect/Handshake/HandshakePasses.h.inc"
