+++ lib/Dialect/Handshake/HandshakeDialect.cpp
@@ -11,6 +11,7 @@
 //===----------------------------------------------------------------------===//
 
 #include "circt/Dialect/Handshake/HandshakeDialect.h"
+#include "circt/Dialect/ESI/ESIDialect.h"
 #include "circt/Dialect/Handshake/HandshakeOps.h"
 #include "circt/Dialect/Seq/SeqDialect.h"
 #include "mlir/IR/Builders.h"
