+++ tools/handshake-runner/CMakeLists.txt
@@ -2,6 +2,7 @@ add_llvm_executable(handshake-runner handshake-runner.cpp Simulation.cpp)
 
 llvm_update_compile_flags(handshake-runner)
 target_link_libraries(handshake-runner PRIVATE
+  CIRCTESI
   CIRCTHandshake
   CIRCTCFToHandshake
   MLIRArithDialect
